CODE_FOH/frontend/requirements.txt
appdirs==1.4.4
appnope==0.1.4
APScheduler==3.11.0
asttokens==2.4.1
async-timeout==5.0.1
beautifulsoup4==4.12.3
blinker==1.8.2
cachelib==0.13.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
crontab==1.0.1
cs50==9.3.4
debugpy==1.8.5
decorator==5.1.1
dlib==19.24.6
exceptiongroup==1.2.2
executing==2.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
Flask==3.0.3
Flask-APScheduler==1.13.1
Flask-Cors==5.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
freezegun==1.5.1
frozendict==2.4.5
greenlet==3.1.1
gunicorn @ git+https://github.com/benoitc/gunicorn.git@903792f152af6a27033d458020923cb2bcb11459
html5lib==1.1
idna==3.10
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.3
jupyter_core==5.7.2
lxml==5.3.0
MarkupSafe==3.0.0
matplotlib-inline==0.1.7
msgspec==0.18.6
multitasking==0.0.11
nest-asyncio==1.6.0
numpy==2.1.2
opencv-python==4.10.0.84
packaging==24.1
pandas==2.2.3
parso==0.8.4
peewee==3.17.6
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
prompt_toolkit==3.0.47
psutil==6.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pyzmq==26.2.0
redis==5.2.0
requests==2.31.0
rq==2.0.0
rq-scheduler==0.14.0
six==1.16.0
soupsieve==2.6
SQLAlchemy==2.0.35
sqlparse==0.5.1
stack-data==0.6.3
termcolor==2.5.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.2
yfinance==0.2.38